Retry joins properly; lose universal timeout logic.
The demo was trying to join a node which during initial lookup of its successor
encountered node 105, which the demo already killed, so when the lookup
protocol failed the join event failed altogether. Now it retries correctly so
that when its known node finds out 105 is dead the joining protocol can pick
back up and finish correctly.

I also got rid of mark_node_dead, since all the logic for doing that really
belongs to specific events and varies based on that. The way it was being done
constituted a variation from Zave's spec.